+++ lib/DependencyScan/CMakeLists.txt
@@ -3,6 +3,7 @@
 add_swift_host_library(swiftDependencyScan STATIC
   DependencyScanningTool.cpp
   ModuleDependencyCacheSerialization.cpp
+  ModuleDependencyScanner.cpp  
   ScanDependencies.cpp
   StringUtils.cpp)
 
@@ -11,4 +12,5 @@ target_link_libraries(swiftDependencyScan INTERFACE
 
 target_link_libraries(swiftDependencyScan PRIVATE
     swiftClangImporter
-    swiftFrontend)
+    swiftAST
+    swiftSerialization)
